Replace previous websocket chat with polling chat
