+++ Templates/chat.html
@@ -11,7 +11,7 @@
 <!--begin::Content-->
 <div class="container-fluid p-0">
     <div class="col p-0">
-        <div id="chat-message-area" class="px-20 pt-10 pb-3"></div>
+        <div id="chat-message-area" class="px-20 pt-10 pb-3" style="background-color: brown;"></div>
         <div class="input-message-area fixed-bottom p-20">
             <div id="typing-status-area"></div>
             <div class="input-group">
