--- chat/routing.py
@@ -1,7 +1,0 @@
-from django.urls import re_path
-
-from . import consumers
-
-websocket_urlpatterns = [
-    #re_path(r"ws/chat/(?P<room_name>\w+)/$", consumers.ChatConsumer.as_asgi())
-]
