+++ chat/urls.py
@@ -3,8 +3,7 @@ from chat import views
 
 app_name = 'chat'
 urlpatterns = [
-    #path('<uuid:room_name>/', views.chat, name='room'),
-    path('ajax/<str:room_name>/', views.chat_ajax, name='room_ajax'),
+    path('<uuid:room_name>/', views.chat, name='room'),
     path('send-message/', views.send_message, name='send_message'),
     path('update/', views.update, name='update')
 ]
